moved wall to game folder and created static-shape-actor
